move OralHistoryBiographicalDisplay to OralHistory::BiographicalComponent
