+++ app/components/work_oh_audio_show_component.html.erb
@@ -153,7 +153,7 @@
               <h3 class="h4"><%= bio.name %></h3>
             <% end %>
 
-            <%= OralHistoryBiographicalDisplay.new(bio).display %>
+            <%= render OralHistory::BiographicalComponent.new(bio) %>
           <% end %>
 
           <%= render "citation", work: work  %>
